move middleware into its own mod
